+++ internal/pkg/clients/types.go
@@ -69,6 +69,8 @@ func (c *Client) setConsensusImage(image string) {
 		c.Image = valueOrDefault(image, configs.ClientImages.Consensus.Teku.String())
 	case "lodestar":
 		c.Image = valueOrDefault(image, configs.ClientImages.Consensus.Lodestar.String())
+	case "grandine":
+		c.Image = valueOrDefault(image, configs.ClientImages.Consensus.Grandine.String())
 	case "nimbus":
 		c.Image = valueOrDefault(image, configs.ClientImages.Consensus.Nimbus.String())
 	}